refactor(cleanup): extract adapter probe helpers to adapter-utils (pass 7)
Moved firstNonEmptyLine, commandLooksLike, and summarizeStatus from
13 adapter files into packages/adapter-utils/src/probe-helpers.ts. All
six local-cli adapters (claude, codex, cursor, gemini, opencode, pi)
plus openclaw-gateway now import the helpers from @paperclipai/adapter-utils
instead of duplicating them.

Also deleted gemini-local/src/server/utils.ts — after the extraction it
was a single re-export, and its only importer (execute.ts) now imports
from @paperclipai/adapter-utils directly.

Removes ~109 lines of byte-for-byte duplication (per .cleanup-reports/
01-dedup.md HIGH #2). Typecheck: all 20 workspace packages green.

Also folds in the pnpm-lock.yaml catchup missed in pass 5 (the
@paperclipai/shared workspace:* dep added to adapter-utils).
