+++ packages/adapter-utils/src/index.ts
@@ -54,3 +54,4 @@ export {
   redactTranscriptEntryPaths,
 } from "./log-redaction.js";
 export { inferOpenAiCompatibleBiller } from "./billing.js";
+export { firstNonEmptyLine, commandLooksLike, summarizeStatus } from "./probe-helpers.js";
