+++ packages/adapter-utils/src/probe-helpers.ts
@@ -1,0 +1,27 @@
+// Shared helpers used by adapter testEnvironment() and execute() paths:
+// output probing, command matching, and status aggregation.
+
+import path from "node:path";
+import type { AdapterEnvironmentCheck, AdapterEnvironmentTestResult } from "./types.js";
+
+export function firstNonEmptyLine(text: string): string {
+  return (
+    text
+      .split(/\r?\n/)
+      .map((line) => line.trim())
+      .find(Boolean) ?? ""
+  );
+}
+
+export function commandLooksLike(command: string, expected: string): boolean {
+  const base = path.basename(command).toLowerCase();
+  return base === expected || base === `${expected}.cmd` || base === `${expected}.exe`;
+}
+
+export function summarizeStatus(
+  checks: AdapterEnvironmentCheck[],
+): AdapterEnvironmentTestResult["status"] {
+  if (checks.some((check) => check.level === "error")) return "fail";
+  if (checks.some((check) => check.level === "warn")) return "warn";
+  return "pass";
+}
