--- packages/adapters/gemini-local/src/server/utils.ts
@@ -1,8 +1,0 @@
-export function firstNonEmptyLine(text: string): string {
-    return (
-        text
-            .split(/\r?\n/)
-            .map((line) => line.trim())
-            .find(Boolean) ?? ""
-    );
-}
