+++ packages/adapters/opencode-local/src/server/models.ts
@@ -1,6 +1,7 @@
 import { createHash } from "node:crypto";
 import os from "node:os";
 import type { AdapterModel } from "@paperclipai/adapter-utils";
+import { firstNonEmptyLine } from "@paperclipai/adapter-utils";
 import {
   asString,
   ensurePathInEnv,
@@ -38,15 +39,6 @@ function dedupeModels(models: AdapterModel[]): AdapterModel[] {
 function sortModels(models: AdapterModel[]): AdapterModel[] {
   return [...models].sort((a, b) =>
     a.id.localeCompare(b.id, "en", { numeric: true, sensitivity: "base" }),
-  );
-}
-
-function firstNonEmptyLine(text: string): string {
-  return (
-    text
-      .split(/\r?\n/)
-      .map((line) => line.trim())
-      .find(Boolean) ?? ""
   );
 }
